fix: Fix Server Action with Client Component issue
