+++ ui/package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "format": "biome format src/**/*{.ts,.tsx} --write"
+    "format": "biome format src/**/{*.ts,*.tsx} --write"
   },
   "dependencies": {
     "@buf/kevinmichaelchen_temporalapis.connectrpc_es": "latest",
